refactor: Outsource config-array mapper
