+++ source/win-main/store/config-to-array.ts
@@ -1,0 +1,35 @@
+export default function configToArrayMapper (config: any): any {
+  // Heh, you're creating an object and call it array? Yes, sort of. What we get
+  // from the config provider is a nested object, because that saves some typing
+  // work (we can just define an object "display" which serves as a prefix for
+  // all keys within it). What this function does is take any nested object and
+  // basically transform it into a flat object with string keys (so, actually,
+  // an associative array, but as we don't have that in JavaScript ... you know
+  // the drill.) where the string keys are the dot-joined key-prefixes.
+  const arr: any = {}
+
+  for (const key of Object.keys(config)) {
+    const value = config[key]
+    const isArray = Array.isArray(value)
+    const isUndefined = value === undefined
+    const isNull = value === null
+    const isNum = typeof value === 'number'
+    const isString = typeof value === 'string'
+    const isBool = typeof value === 'boolean'
+    if (isArray || isUndefined || isNull || isNum || isString || isBool) {
+      // Yep, above are all (possibly) checks we have to perform in order to be
+      // certain that value is not a "normal" object with key-value pairs. This
+      // means we have reached one leaf and can begin traversing up
+      arr[key] = value
+    } else {
+      // Traverse one level deeper.
+      const mapped = configToArrayMapper(value)
+      for (const mappedKey in mapped) {
+        // Add the namespace here
+        arr[key + '.' + mappedKey] = mapped[mappedKey]
+      }
+    }
+  }
+
+  return arr
+}
